+++ test/integration/DecimalsTests.t.sol
@@ -121,12 +121,12 @@ contract DecimalsTests is BaseTest {
     }
 
     function test_fuzz_node_swing_price_deposit_never_exceeds_max_6decimals(
-        uint256 maxSwingFactor,
+        uint64 maxSwingFactor,
         uint256 targetReserveRatio,
         uint256 seedAmount,
         uint256 depositAmount
     ) public {
-        maxSwingFactor = bound(maxSwingFactor, 0.01 ether, 0.99 ether);
+        maxSwingFactor = uint64(bound(maxSwingFactor, 0.01 ether, 0.99 ether));
         targetReserveRatio = bound(targetReserveRatio, 0.01 ether, 0.99 ether);
         seedAmount = bound(seedAmount, 1 ether, maxDeposit);
         depositAmount = bound(depositAmount, 1 ether, maxDeposit);
@@ -182,12 +182,12 @@ contract DecimalsTests is BaseTest {
     }
 
     function test_fuzz_node_swing_price_redeem_never_exceeds_max_6decimals(
-        uint256 maxSwingFactor,
+        uint64 maxSwingFactor,
         uint256 targetReserveRatio,
         uint256 seedAmount,
         uint256 withdrawalAmount
     ) public {
-        maxSwingFactor = bound(maxSwingFactor, 0.01 ether, 0.99 ether);
+        maxSwingFactor = uint64(bound(maxSwingFactor, 0.01 ether, 0.99 ether));
         targetReserveRatio = bound(targetReserveRatio, 0.01 ether, 0.99 ether);
         seedAmount = bound(seedAmount, 100 ether, 1e36);
